requirements.txt
SpeechRecognition
pyaudio
pyautogui
pygetwindow
keyboard
fastapi
faster_whisper
uvicorn
pydantic
uvicorn
httpx
aiohttp
langchain_core
langgraph
Pillow
dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib